// src/shared/api/models/getApiGiftCardsOnlineType200ItemsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiGiftCardsOnlineType200ItemsItem = {
    id?: string;
    name?: string;
    /** @nullable */
    description?: string | null;
    price?: number;
    expiryDays?: number;
    sellOnline?: boolean;
    createdAt?: string;
    updatedAt?: string;
};